requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
Flask==3.1.1
frozenlist==1.7.0
googlesearch-python==1.3.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instaloader==4.11
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.6.3
mutagen==1.47.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pycryptodomex==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
pyTelegramBotAPI==4.27.0
pytesseract==0.3.13
python-telegram-bot==21.0.1
pytube==15.0.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
Werkzeug==3.1.3
yarl==1.20.1
yt-dlp==2024.5.27
